[dyn-cache] divide storage module so i can understand it
